% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_custom_output.R
\name{DECEPTICON__custom_output}
\alias{DECEPTICON__custom_output}
\title{Integrating analysis strategies and outputting DECEPTICON results}
\usage{
DECEPTICON__custom_output(path, signature.matrix)
}
\arguments{
\item{path}{The storage path for the results of the analysis strategies}

\item{signature.matrix}{a i*l matrix with i genes and l cell types.}
}
\description{
Integrating analysis strategies and outputting DECEPTICON results
}
