% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_method.R
\name{DECEPTICON_batman}
\alias{DECEPTICON_batman}
\title{Deconvolute using Batman}
\usage{
DECEPTICON_batman()
}
\arguments{
\item{data}{a m*n matrix with m genes and n samples}
}
\description{
Deconvolute using Batman
}
