% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_custom_methods.R
\name{DECEPTICON_batman_custom}
\alias{DECEPTICON_batman_custom}
\title{Deconvolute using Batman and custom signature matrix}
\usage{
DECEPTICON_batman_custom(signature.matrix)
}
\arguments{
\item{signature.matrix}{a i*l matrix with i genes and l cell types.}

\item{data}{a m*n matrix with m genes and n samples}
}
\description{
Deconvolute using Batman and custom signature matrix
}
