% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_method.R
\name{DECEPTICON_bseq}
\alias{DECEPTICON_bseq}
\title{Deconvolute using Bseq-SC}
\usage{
DECEPTICON_bseq(data)
}
\arguments{
\item{data}{a m*n matrix with m genes and n samples}
}
\description{
Deconvolute using Bseq-SC
}
