% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_method.R
\name{DECEPTICON_ciber}
\alias{DECEPTICON_ciber}
\title{Deconvolute using CIBERSORT(Robust enumeration of cell subsets from tissue expression profiles)}
\usage{
DECEPTICON_ciber(mixture_file, light)
}
\arguments{
\item{light}{set to TRUE for light version of DECEPTICON}

\item{mix_file}{a m*n matrix with m genes and n samples}
}
\description{
Deconvolute using CIBERSORT(Robust enumeration of cell subsets from tissue expression profiles)
}
