% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_method.R
\name{DECEPTICON_ciber_abs}
\alias{DECEPTICON_ciber_abs}
\title{Deconvolute using CIBERSORT-abs}
\usage{
DECEPTICON_ciber_abs(mixture_file, light)
}
\arguments{
\item{light}{set to TRUE for light version of DECEPTICON}

\item{mix_file}{a m*n matrix with m genes and n samples}
}
\description{
Deconvolute using CIBERSORT-abs
}
