% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_custom_methods.R
\name{DECEPTICON_ciber_abs_custom}
\alias{DECEPTICON_ciber_abs_custom}
\title{Deconvolute using CIBERSORT-abs and custom signature matrix}
\usage{
DECEPTICON_ciber_abs_custom(mixture_file, signature.matrix)
}
\arguments{
\item{mixture_file}{a m*n matrix with m genes and n samples}

\item{signature.matrix}{a i*l matrix with i genes and l cell types.}
}
\description{
Deconvolute using CIBERSORT-abs and custom signature matrix
}
