% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_custom_methods.R
\name{DECEPTICON_custom_methods}
\alias{DECEPTICON_custom_methods}
\title{deconvolution for cell type interpretation by consensus}
\usage{
DECEPTICON_custom_methods(signature.matrix, bulk.samples, RUNpath)
}
\arguments{
\item{signature.matrix}{a i*l matrix with i genes and l cell types.}

\item{bulk.samples}{a m*n matrix with m genes and n samples}

\item{RUNpath}{Working path for storing files}
}
\description{
deconvolution for cell type interpretation by consensus
}
