% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_method.R
\name{DECEPTICON_decon}
\alias{DECEPTICON_decon}
\title{Deconvolute using DeconRNAseq(DeconRNASeq: a statistical framework for deconvolution of heterogeneous tissue samples based on mRNA-Seq data)}
\usage{
DECEPTICON_decon(dataset)
}
\arguments{
\item{dataset}{a m*n matrix with m genes and n samples}
}
\description{
Deconvolute using DeconRNAseq(DeconRNASeq: a statistical framework for deconvolution of heterogeneous tissue samples based on mRNA-Seq data)
}
