% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_custom_methods.R
\name{DECEPTICON_decon_custom}
\alias{DECEPTICON_decon_custom}
\title{Deconvolute using DeconRNAseq and custom signature matrix}
\usage{
DECEPTICON_decon_custom(dataset, signature.matrix)
}
\arguments{
\item{dataset}{a m*n matrix with m genes and n samples}

\item{signature.matrix}{a i*l matrix with i genes and l cell types.}
}
\description{
Deconvolute using DeconRNAseq and custom signature matrix
}
