% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_method.R
\name{DECEPTICON_epic}
\alias{DECEPTICON_epic}
\title{Deconvolute using EPIC(Simultaneous enumeration of cancer an immune cell types from bulk tumor gene expression data)}
\usage{
DECEPTICON_epic(mix.mat)
}
\arguments{
\item{mix.mat}{a m*n matrix with m genes and n samples}
}
\description{
Deconvolute using EPIC(Simultaneous enumeration of cancer an immune cell types from bulk tumor gene expression data)
}
