% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_custom_methods.R
\name{DECEPTICON_epic_custom}
\alias{DECEPTICON_epic_custom}
\title{Deconvolute using EPIC and custom signature matrix}
\usage{
DECEPTICON_epic_custom(mix.mat, signature.matrix)
}
\arguments{
\item{mix.mat}{a m*n matrix with m genes and n samples}

\item{signature.matrix}{a i*l matrix with i genes and l cell types.}
}
\description{
Deconvolute using EPIC and custom signature matrix
}
