% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_method.R
\name{DECEPTICON_mcp}
\alias{DECEPTICON_mcp}
\title{Deconvolute using MCPcounter(Estimating the population abundance of tissue-infiltrating immune and stromal cell populations using gene expression)}
\usage{
DECEPTICON_mcp(data)
}
\arguments{
\item{data}{a m*n matrix with m genes and n samples}
}
\description{
Deconvolute using MCPcounter(Estimating the population abundance of tissue-infiltrating immune and stromal cell populations using gene expression)
}
