% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_custom_methods.R
\name{DECEPTICON_mcp_custom}
\alias{DECEPTICON_mcp_custom}
\title{Deconvolute using MCPcounter and custom signature matrix}
\usage{
DECEPTICON_mcp_custom(data, signature.matrix)
}
\arguments{
\item{data}{a m*n matrix with m genes and n samples}

\item{signature.matrix}{a i*l matrix with i genes and l cell types.}
}
\description{
Deconvolute using MCPcounter and custom signature matrix
}
