% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_method.R
\name{DECEPTICON_methods}
\alias{DECEPTICON_methods}
\title{deconvolution for cell type interpretation by consensus}
\usage{
DECEPTICON_methods(bulk.samples, RUNpath, light)
}
\arguments{
\item{bulk.samples}{a m*n matrix with m genes and n samples}

\item{RUNpath}{Working path for storing files}

\item{light}{set to TRUE for light version of DECEPTICON}
}
\description{
deconvolution for cell type interpretation by consensus
}
