% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_method.R
\name{DECEPTICON_music}
\alias{DECEPTICON_music}
\title{Deconvolute using MuSic}
\usage{
DECEPTICON_music(data)
}
\arguments{
\item{data}{a m*n matrix with m genes and n samples}
}
\description{
Deconvolute using MuSic
}
