% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_output.R
\name{DECEPTICON_output}
\alias{DECEPTICON_output}
\title{Integrating analysis strategies and outputting DECEPTICON results}
\usage{
DECEPTICON_output(path, light = FALSE)
}
\arguments{
\item{path}{The storage path for the results of the analysis strategies}
}
\description{
Integrating analysis strategies and outputting DECEPTICON results
}
