% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_method.R
\name{DECEPTICON_quan}
\alias{DECEPTICON_quan}
\title{Deconvolute using quanTIseq(Molecular and pharmacological modulators of the tumor immune contexture revealed by deconvolution of RNA-seq data)}
\usage{
DECEPTICON_quan(mix.mat)
}
\arguments{
\item{mix.mat}{a m*n matrix with m genes and n samples}
}
\description{
Deconvolute using quanTIseq(Molecular and pharmacological modulators of the tumor immune contexture revealed by deconvolution of RNA-seq data)
}
