% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_method.R
\name{DECEPTICON_scdc}
\alias{DECEPTICON_scdc}
\title{Deconvolute using SCDC}
\usage{
DECEPTICON_scdc(data)
}
\arguments{
\item{data}{a m*n matrix with m genes and n samples}
}
\description{
Deconvolute using SCDC
}
