% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_output.R
\name{optimal_id}
\alias{optimal_id}
\title{Select the analysis strategies with the highest correlation}
\usage{
optimal_id(matrix, n)
}
\arguments{
\item{matrix}{a m*m correlation matrix with m analysis strategies}

\item{n}{The top n groups of analysis strategies with the highest correlation}
}
\description{
Select the analysis strategies with the highest correlation
}
