% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_output.R
\name{run_DECEPTICON}
\alias{run_DECEPTICON}
\title{DEconvolution for CEll Proportion esTImation by CONsensus}
\usage{
run_DECEPTICON(
  bulk.samples,
  RUNpath,
  path,
  light = FALSE,
  custom.signature = FALSE,
  signature.matrix = NULL
)
}
\arguments{
\item{bulk.samples}{a m*n matrix with m genes and n samples}

\item{RUNpath}{Working path for storing files}

\item{path}{The storage path for the results of the analysis strategies}

\item{light}{set to TRUE for light version of DECEPTICON}

\item{custom.signature}{set to TRUE for custom signature matrix}

\item{signature.matrix}{a i*l matrix with i genes and l cell types.}
}
\description{
DEconvolution for CEll Proportion esTImation by CONsensus
}
