% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICON_output.R
\name{substrRight}
\alias{substrRight}
\title{Determine whether the expression template of the analysis strategies is the same}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{The name of the analysis strategies}

\item{n}{The last n digits of the  name of the analysis strategies}
}
\description{
Determine whether the expression template of the analysis strategies is the same
}
